+++ src/index.ts
@@ -5,3 +5,4 @@ export * from "./accessor";
 export * from "./controlled";
 export * from "./group";
 export * from "./antd";
+export * from "./state";